refactor: reorganize API endpoint management components
- Removed the old ApiEndpointDialog implementation and replaced it with a new structure that includes ApiEndpointForm, ApiEndpointList, and ApiEndpointListItem components for improved modularity.
- Enhanced the API endpoint management experience by integrating new components into the dialog.
- Updated the handling of API endpoint configurations to streamline the addition, editing, and deletion processes.
- Improved code organization and maintainability by separating concerns into distinct components.
